# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MYSIG API",
        "description": "MYSIG API for Frontend",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
...
        "/api/policy/showTravelSubid/{language_type}": {
            "get": {
                "tags": [
                    "Travel"
                ],
                "summary": " showTravelSubid",
                "description": "Show Home list API",
                "operationId": "showTravelSubid",
                "parameters": [
                    {
                        "name": "language_type",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/policy/showTravelZone": {
            "post": {
                "tags": [
                    "Travel"
                ],
                "summary": " showTravelZone",
                "description": "Show travel list API",
                "operationId": "showTravelZone",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "country_type_id": {
                                        "type": "string"
                                    },
                                    "language_type": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/policy/showTravelList": {
            "post": {
                "tags": [
                    "Travel"
                ],
                "summary": " ShowTravelList",
                "description": "Show travel list API",
                "operationId": "ShowTravelList",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "choose_contury_type": {
                                        "type": "string"
                                    },
                                    "for_age": {
                                        "type": "string"
                                    },
                                    "for_zone": {
                                        "type": "string"
                                    },
                                    "for_validity": {
                                        "type": "string"
                                    },
                                    "start_date": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "JWT bearer",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}